Extrais le tableau des services dans un module
